+++ src/main/java/com/company/calculator/services/ArithmeticCalculatorSubtractServiceImpl.java
@@ -1,11 +1,11 @@
 package com.company.calculator.services;
 
-import java.math.BigInteger;
+import java.math.BigDecimal;
 
 public class ArithmeticCalculatorSubtractServiceImpl implements ArithmeticCalculatorSubtractService {
     @Override
-    public BigInteger subtract(final BigInteger firstTerm,
-                               final BigInteger secondTerm) {
+    public BigDecimal subtract(final BigDecimal firstTerm,
+                               final BigDecimal secondTerm) {
         return firstTerm.subtract(secondTerm);
     }
 }
